Seperate attempt_guess into two methods
Previously the attempt_guess method was serving two purpose which
was validating the guess and then changing the game state on the
basis of correct or incorrect guess

Now there are two methods , one is validate_guess whose result
will affect the execution of attempt_guess. As long as guess is not valid,
game cannot go further.